# Switch to string path traverse instead of array