Massive perf improvement in synthetic benchmark:

Before:
```
# GitHub Actions
simple x 10k: 2.993s
wildcard x 10k: 11.394s
globstar x 100: 11.934s
```

After 🤯
```
# GitHub Actions
simple x 10k: 926.962ms
wildcard x 10k: 3.439s
globstar x 100: 2.562s
```